pump java-client version && some fix (#48)

+++ app/src/main/java/io/hstream/tools/ReadBench.java
@@ -215,7 +215,7 @@ public class ReadBench {
     Utils.BufferedProducerOpts batchProducerOpts = new Utils.BufferedProducerOpts();
 
     @CommandLine.Option(names = "--rate-limit")
-    int rateLimit = 100000;
+    long rateLimit = 100000;
 
     @CommandLine.ArgGroup(exclusive = false)
     Utils.PayloadOpts payloadOpts = new Utils.PayloadOpts();
