+++ app/src/main/java/io/hstream/tools/WriteBench.java
@@ -315,7 +315,7 @@ public class WriteBench {
     Utils.BufferedProducerOpts batchProducerOpts = new Utils.BufferedProducerOpts();
 
     @CommandLine.Option(names = "--rate-limit")
-    int rateLimit = 100000;
+    long rateLimit = 100000;
 
     @CommandLine.ArgGroup(exclusive = false)
     Utils.PayloadOpts payloadOpts = new Utils.PayloadOpts();
